+++ inst/htmlwidgets/morrisjs.js
@@ -12,13 +12,18 @@ HTMLWidgets.widget({
   },
 
   renderValue: function(el, x, instance) {
-    // Workaround if only one series is plotted
+    
+  // Workaround if only one series is plotted
+  if ("ykeys" in x){
     if (x.ykeys.constructor !== Array) {
       x.ykeys = new Array(x.ykeys);
     }
     if (x.labels.constructor !== Array) {
       x.labels = new Array(x.labels);
     }
+  }
+    
+    
   
   // Create the graph  
   var mjs;
@@ -31,6 +36,9 @@ HTMLWidgets.widget({
       break;
     case "Bar": 
       mjs = new Morris.Bar(x);
+      break;
+    case "Donut":
+      mjs = new Morris.Donut(x);
       break;
   }